Consegui colocar o botao "EDITAR", para funcionar, agora todos os botoes das telas Cadastro cliente, produto e usuario, estão funcionando. tela login esta funcionando tambem...

+++ src/model/BensProduto.java
@@ -17,7 +17,7 @@ public class BensProduto {
     private int ID_PRODUTO;
     private String Descricao;
     private double Valor;
-    private int Total;
+    private double Total;
     private String Pesquisa;
 
     public BensProduto() {
@@ -64,7 +64,7 @@ public class BensProduto {
         return Total;
     }
 
-    public void setTotal(int Total) {
+    public void setTotal(double Total) {
         this.Total = Total;
     }
   public String getPesquisa() {
